refactor: move CiContext to homeboy-extension-contract
CiContext is a pure serde struct (0 impls, no custom serde helpers) whose field
types CiJobMapping and CiLocalContext already live in the contract crate (from
the manifest-config cuts). It's a field dependency of the aggregate
TestCommandOutput/LintCommandOutput result clusters.

Moves it out of ci_profile.rs into the contract crate; the CI-profile resolution
functions stay in core. Core re-exports it so all crate::extension paths stay
stable. This + AppliedRefactor (already in homeboy-refactor-contract) clears the
non-test-specific field deps of the top-level test/lint result clusters.
